separate hook for fetching from the server and in place actions
